[FLINK-16577][table-planner-blink] Fix numeric type mismatch error in  column interval relmetadata
This closes #12326

+++ flink-table/flink-table-planner-blink/src/test/java/org/apache/flink/table/planner/catalog/CatalogStatisticsTest.java
@@ -55,6 +55,7 @@ import org.apache.calcite.util.ImmutableBitSet;
 import org.junit.Before;
 import org.junit.Test;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
@@ -158,7 +159,9 @@ public class CatalogStatisticsTest {
 		// long type
 		assertEquals(46.0, mq.getDistinctRowCount(t1, ImmutableBitSet.of(0), null), 0.0);
 		assertEquals(154.0, mq.getColumnNullCount(t1, 0), 0.0);
-		assertEquals(ValueInterval$.MODULE$.apply(-123L, 763322L, true, true), mq.getColumnInterval(t1, 0));
+		assertEquals(
+				ValueInterval$.MODULE$.apply(BigDecimal.valueOf(-123L), BigDecimal.valueOf(763322L), true, true),
+				mq.getColumnInterval(t1, 0));
 
 		// string type
 		assertEquals(40.0, mq.getDistinctRowCount(t1, ImmutableBitSet.of(1), null), 0.0);
@@ -183,7 +186,9 @@ public class CatalogStatisticsTest {
 		// long type
 		assertEquals(46.0, mq.getDistinctRowCount(t1, ImmutableBitSet.of(0), null), 0.0);
 		assertEquals(154.0, mq.getColumnNullCount(t1, 0), 0.0);
-		assertEquals(ValueInterval$.MODULE$.apply(-123L, 763322L, true, true), mq.getColumnInterval(t1, 0));
+		assertEquals(
+				ValueInterval$.MODULE$.apply(BigDecimal.valueOf(-123L), BigDecimal.valueOf(763322L), true, true),
+				mq.getColumnInterval(t1, 0));
 
 		// string type
 		assertEquals(40.0, mq.getDistinctRowCount(t1, ImmutableBitSet.of(1), null), 0.0);
@@ -324,7 +329,9 @@ public class CatalogStatisticsTest {
 		// long type
 		assertEquals(23.0, mq.getDistinctRowCount(rel, ImmutableBitSet.of(1), null), 0.0);
 		assertEquals(77.0, mq.getColumnNullCount(rel, 1), 0.0);
-		assertEquals(ValueInterval$.MODULE$.apply(-123L, 763322L, true, true), mq.getColumnInterval(rel, 1));
+		assertEquals(
+				ValueInterval$.MODULE$.apply(BigDecimal.valueOf(-123L), BigDecimal.valueOf(763322L), true, true),
+				mq.getColumnInterval(rel, 1));
 
 		// string type
 		assertEquals(20.0, mq.getDistinctRowCount(rel, ImmutableBitSet.of(2), null), 0.0);
@@ -343,7 +350,9 @@ public class CatalogStatisticsTest {
 		// double type
 		assertEquals(73.0, mq.getDistinctRowCount(rel, ImmutableBitSet.of(4), null), 0.0);
 		assertEquals(27.0, mq.getColumnNullCount(rel, 4), 0.0);
-		assertEquals(ValueInterval$.MODULE$.apply(-123.35, 7633.22, true, true), mq.getColumnInterval(rel, 4));
+		assertEquals(
+				ValueInterval$.MODULE$.apply(BigDecimal.valueOf(-123.35), BigDecimal.valueOf(7633.22), true, true),
+				mq.getColumnInterval(rel, 4));
 	}
 
 	private void alterTableStatisticsWithUnknownRowCount(
